+++ src/components/tradeui/CHANGELOG.md
@@ -1,3 +1,5 @@
+# Jan. 3rd, 2022 mini-update #4
+- Add reset / default buttons to Control Panel
 # Jan. 3rd, 2022 mini-update #3
 - Add settings version keys
 - Remove CORS backend service from prod :)